Creating Minecraft cog Figured it's about time it gets its own space for commands and such. Moved all Minecraft related commands inside it. Also added a +locations command which links to the Address Book.
- Also fixed various bugs:
- Order of operations error in +data command. it was checking lapislord 
file, when it didn't exist yet. Made +data owner only for the time 
being.
- Forgot to await the shutdown command, now fixed.
